Merged PR 86183: Orthographic view attachments render directly into sheet view
Where "orthographic" includes both 2d and 3d view attachments.
If the attached view has a perspective transform, we do not draw it. I have yet to come across such a view attachment in the real world - will circle back later to reimplement.
Previously all 3d attachments were rendered off-screen and displayed as raster tiles in the sheet.
Now you can interact directly with the elements in the attached view, whether 2d or 3d, with synchronized selection etc.
Previously, for 2d views we would select+draw the attached view's tiles through a transform. This bypassed some logic, potentially producing different results than would be seen if you opened the attached view directly.
Now, instead, we draw the view (2d or 3d) into a lightweight off-screen viewport with a barebones RenderTarget, collect the resultant graphics, and add them to the sheet view's scene through a transform.
Previously we had a hard-coded z range for 2d frustum, [-1..1]. Now, we expand the sheet view's frustum backwards to accommodate the depth of the 3d attachments. Display priority values continue to map to [-1..1]. The near plane of each attachment's frustum is aligned with the sheet view based on display priority.
When we draw each view attachment, we compute the intersection of the attached view's frustum with the sheet view's frustum, for culling during tile selection; and adjust the size of its ViewRect based on pixel size, to ensure tiles of appropriate LOD are selected.

As a result of this PR, the "clip mask" shaders are no longer used. We should probably delete them.

Related work items: #333784, #333822