Merged PR 90454: Add `.env` template and General Purpose Backend support to ui-test-app.  Various other cleanup
Update the ui-test-app to have a template `.env` file at the root of the app to help in letting people know the different configuration options.

Update the backend to use the `.env` file as well, previously it would only pick up the file if it existed in the `imodeljs-config` repo and not in the `process.env()`.