Merged PR 90996: Introduce the @bentley/physical-material-backend package
Changes:

- Introduce the `@bentley/physical-material-backend` package
- Add `PhysicalMaterial.createCode`
- Add `PhysicalMaterial.create`

The goal is also to set up a standard pattern for domain handlers associated with *common layer* BIS schemas. The expectation is that many applications would use these domains, so being part of the imodeljs monorepo (at least for now) seems to make sense. However, we want things to be more modular than the previous strategy of just lumping them into `@bentley/imodeljs-backend` and `@bentley/imodeljs-common`. Had this strategy been in place earlier, we would not have created the `core/backend/src/domains` and `core/common/src/domains` directories.

> Note: The PhysicalMaterial schema happens not to need any *props* interfaces, so there is not a corresponding `@bentley/physical-material-common` package as there will be for most domains.

As part of this pattern, we also want to finalize our strategy for publishing BIS schemas as npm packages.