Merged PR 94433: Fix 368624: Shader compilation error in ubuntu 20.04
Surface vertex shader computes v_eyeSpace based on u_frustum.
Fragment shader uses that to write fragment depth, if log Z enabled; and uses it for surface discard.
If the shader is a hilite shader, **and** log Z is not enabled, the varying is unused.
Clever compilers (chiefly linux) correctly optimize out the varying and the uniform in this case; we throw an exception when getProgramLocation() fails to find the uniform.

Omit this logic from hilite shaders if log Z disabled.

Related work items: #368624