Merged PR 101868: Move default tile options into imodeljs-common to help tile agent stay in sync w/ frontends
If the tile-agent is generating tiles with different flags than a frontend wants, the pregenerated tiles do us no good.  Since most frontends always use the default TileAdmin configuration, we can make it easy for the tile-agent to stay in sync with those default options without having to pull in the frontend package.