Merged PR 104774: Cleaned up ViewAttributes menu a little
Cleaned up ViewAttributes menu a little.  Standardized the use of hr elements (divider lines) and used some bolding & color changes to make the different sections stand out more so that they are easier to find.

Also fixed a bug with selecting the Ground Plane which (when neither the ground plane or sky box were selected) selected the skybox instead.

Made several spacing adjustments to make things more readable.

Made Axis, Sun Direction, and Sensor X, Y, and Z values inline to reduce the length of the thematic menu.

I had tried to double up the check-boxes on the attributes flags (make 2 columns of them) in order to shorten this part of the menu, but wasn't successful in dealing with the Scaling and Color widgets for Monochrome when they became visible.