Merged PR 105277: Make IModelApp globally accessible for debugging.
App teams frequently ask us to debug their apps. More often than not, debuggable source code/maps are unavailable; and, with the exception of one of the Design Review apps, so is key-in command entry.
In `IModelApp.startup`, set `window.iModelAppForDebugger` to `this`. Reset it in `shutdown`. Now we can debug productively.
It would be useful to additionally make an extension that wraps the `frontend-devtools` package so we could load in all of those useful key-ins as needed.