Merged PR 105375: Fix for crash in volume classification
For some reason the classifier geometry is no longer being created with a SurfaceType of VolumeClassifier.  This results in an empty volume classification command list and was causing a crash when trying to flash anything when volume classification is being used.  This PR fixes the crash, but does not fix the underlying problem of the classifier geometry no longer being typed correctly.  Currently everything will be classified as Outside.  I will attempt to fix the other problem in another PR, but this will at least prevent the crash for now.