Merged PR 105575: Fix bugs in locking, deleting of assemblies
Deleting a parent element recursively deletes its children. Accordingly, the parent element's populateRequest method must report that locks are needed on the child elements. This allows the ElementEditor helper class to acquire all needed locks before deleting an assembly.

If an app were to specify both the parent element and the child elements in a single call to delete, the delete logic may find that one of the children has already been deleted (as a consequence of deleting the parent). It should just skip an already-deleted element.

ElementEditor was failing to reserve codes on new elements before inserting them.