Merged PR 106889: Logging error and returning blank when UiCore.translate called without UiCore.initialize being called beforehand
If UiCore.initialize is not called before UiCore.translate, UiCore.translate will log an error and return a blank string as the localized string. Before this change, an error was thrown from UICore.i18n saying "UiCore not initialized".

This change has been introduced because at least one application was using the ui-core package without calling UiCore.initialize and considered adding the use of a localized string within a component that was not previously using a localized string a breaking change because of the thrown error.

This brings up a more basic question. Should the UI packages never be throwing errors when we encounter an error condition? Perhaps we should only log errors and fail quietly.