Merged PR 107505: Add methods to get reality model appearance overrides.
In implementing the appearance overrides for reality models, we noticed that I had omitted a method to get the current appearance override.  This adds that test and also aligns the reality model appearance API more closely with that of regular models.  This PR also continues to refine test for large reality model realignment to avoid it being misapplied.

Related work items: #436677