Merged PR 109180: Never attempt to copy electron in CopyExternalsPlugin
In addition to builtinModules, we need to special-case electron in `CopyExternalsPlugin` as well, since it should _never_ be copied to dist/node_modules.

Electron really acts as a replacement for node, and so when running as such, imports of the "electron" module function more like node builtins.  Also, electron-builder does not allow electron to be a prod dependency anyway, so before we were really telling users to do the exact opposite of what they should have done.

Also fix a minor issue with `MissingExternalWarning` logs sometimes omitting the first line.